requirements.txt
aiohttp==2.0.7
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.3
Django==1.11.1
future==0.16.0
Jinja2==2.9.6
MarkupSafe==1.0
multidict==2.1.5
python-telegram-bot==6.0.1
pytz==2017.2
PyYAML==3.12
render==1.0.0
requests==2.14.2
telepot==12.0
urllib3==1.21.1
yarl==0.10.2
